+++ db/models/Item.go
@@ -3,19 +3,19 @@ package models
 import "time"
 
 type Item struct {
-	ID int64
-	Cat *Category
-	Name string
-	Desc string
-	Preview string
-	Type string
-	BLink string
-	Tags []string
-	Price float64
-	Author string
-	DownCnt int64
-	Scores int64
-	Status string
+	ID         int64
+	Cat        *Category
+	Name       string
+	Desc       string
+	Preview    string
+	Type       string
+	BLink      string
+	Tags       []string
+	Price      float64
+	Author     string
+	DownCnt    int64
+	Scores     float64
+	Status     string
 	CreateTime time.Time
 	UpdateTime time.Time
 }
@@ -25,7 +25,7 @@ func (i *Item) New() *Item {
 	return i
 }
 
-func MakeItem()  *Item{
+func MakeItem() *Item {
 	item := Item{}
 	item.Cat = &Category{}
 	return &item
